Исправил баг с несоответствием размера данных состояния коннектора в сервере и карте

+++ network/include/tcp-client.h
@@ -35,7 +35,7 @@ public:
 
     void sendRequest(StructureType stype, double update_interval = 0.0);
 
-    void sendSwitchState(QString conn_name, int state_fwd, int state_bwd);
+    void sendSwitchState(QString conn_name, int8_t state_fwd, int8_t state_bwd);
 
     void sendSignalState(QString conn_name, int sig_dir, bool open);
